defaulting to 16550 uart for all simulations
resolved Issue#29. uart0 (16550) has become the default uart for all simulations henceforth. This change will reflect in a slight lower dmips value because of code relocation casued by changes in the putchar function. The reduction however will be insignificant for larger iterations of dmips.  it will be good to have a "putchar" independent compile of the benchmarks soon.