reverting to a fully pipelined AXI4 fabric
see issue#5

This version of the fabric is completely pipelined. 

There are no CReg instances to reduce latency. This should significantly help in synthesis since the number of slaves on this AXI have also reduced.

Some fields such as qos, prot, lock, cach and region which are not currently required have also been removed. This reduces the effective width of the read and write channels.